chore: update to typescript
