+++ package.json
@@ -15,5 +15,10 @@
     "react-dom": "17.0.2",
     "remark": "^13.0.0",
     "remark-html": "^13.0.1"
+  },
+  "devDependencies": {
+    "@types/node": "^16.6.0",
+    "@types/react": "^17.0.17",
+    "typescript": "^4.3.5"
   }
 }
